+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## master (unreleased)
 
+### New features
+
+* Add new command `projectile-find-file-in-directory` that allows you to jump to files in any directory
+
 ## 0.9.2 (07/16/2013)
 
 ### New features